requirements.txt
discord==1.7.3
discord.py==1.7.3
python-dotenv==0.21.0
